+++ noodle_WE4b/src/app/app.module.ts
@@ -28,12 +28,17 @@ import { ForumFileAttachmentComponent } from './pages/forums/forum-file-attachme
 import { ForumEmojiPickerComponent } from './pages/forums/forum-emoji-picker/forum-emoji-picker.component';
 
 import { ParticipantsListComponent } from './pages/partipants-ue/participants-list/participants-list.component';
-import { UeRegistrationComponent } from './pages/ue-registration/ue-registration.component';
 import { UeHeaderComponent } from './pages/partipants-ue/ue-header/ue-header.component';
 import { ParticipantsFiltersComponent } from './pages/partipants-ue/participants-filters/participants-filters.component';
 import { ParticipantsGridComponent } from './pages/partipants-ue/participants-grid/participants-grid.component';
 import { ParticipantCardComponent } from './pages/partipants-ue/participant-card/participant-card.component';
 import { ParticipantsEmptyStateComponent } from './pages/partipants-ue/participants-empty-state/participants-empty-state.component';
+import { UeFormComponent } from './pages/ue-registration/ue-registration/components/ue-form/ue-form.component';
+import { UeListComponent } from './pages/ue-registration/ue-registration/components/ue-list/ue-list.component';
+import { UserManagementComponent } from './pages/ue-registration/ue-registration/components/user-management/user-management.component';
+import { UeCardComponent } from './pages/ue-registration/ue-registration/components/ue-card/ue-card.component';
+import { ImageUploadComponent } from './pages/ue-registration/ue-registration/components/image-upload/image-upload.component';
+import { UeRegistrationComponent} from "./pages/ue-registration/ue-registration/ue-registration.component";
 
 @NgModule({
   declarations: [
@@ -48,7 +53,6 @@ import { ParticipantsEmptyStateComponent } from './pages/partipants-ue/participa
     ForumListComponent,
     ForumDetailComponent,
     ParticipantsListComponent,
-    UeRegistrationComponent,
     ForumHeaderComponent,
     ForumCreateCardComponent,
     ForumFiltersCardComponent,
@@ -64,6 +68,12 @@ import { ParticipantsEmptyStateComponent } from './pages/partipants-ue/participa
     ParticipantsGridComponent,
     ParticipantCardComponent,
     ParticipantsEmptyStateComponent,
+    UeFormComponent,
+    UeListComponent,
+    UserManagementComponent,
+    UeCardComponent,
+    ImageUploadComponent,
+    UeRegistrationComponent
   ],
   imports: [
     BrowserModule,
