+++ noodle_WE4b/src/app/pages/ue-registration/ue-registration/components/image-upload/image-upload.component.html
@@ -1,0 +1,1 @@
+<p>image-upload works!</p>
