+++ noodle_WE4b/src/app/pages/ue-registration/ue-registration/components/ue-card/ue-card.component.html
@@ -1,0 +1,1 @@
+<p>ue-card works!</p>
