+++ noodle_WE4b/src/app/pages/ue-registration/ue-registration/components/ue-card/ue-card.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-ue-card',
+  templateUrl: './ue-card.component.html',
+  styleUrls: ['./ue-card.component.css']
+})
+export class UeCardComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}
