+++ noodle_WE4b/src/app/pages/ue-registration/ue-registration/components/ue-list/ue-list.component.html
@@ -1,0 +1,1 @@
+<p>ue-list works!</p>
