+++ noodle_WE4b/src/app/pages/ue-registration/ue-registration/components/ue-list/ue-list.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-ue-list',
+  templateUrl: './ue-list.component.html',
+  styleUrls: ['./ue-list.component.css']
+})
+export class UeListComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}
