+++ noodle_WE4b/src/app/pages/ue-registration/ue-registration/components/user-management/user-management.component.html
@@ -1,0 +1,1 @@
+<p>user-management works!</p>
